chocoq lib fix

+++ janusq/application/chocoq/chocoq/problems/facility_location_problem.py
@@ -18,7 +18,7 @@ class FacilityLocationProblem(LcboModel):
     def get_feasible_solution(self):
         """ 根据约束寻找到一个可行解 """
         import numpy as np
-        fsb_lst = np.zeros(len(self.variables))
+        fsb_lst = np.zeros(len(self.variables), dtype=int)
         fsb_lst[0] = 1
         for i in range(self.num_demands):
             fsb_lst[self.num_facilities + self.num_facilities * i] = 1
