+++ janusq/application/chocoq/chocoq/problems/job_scheduling_problem.py
@@ -17,7 +17,7 @@ class JobSchedulingProblem(LcboModel):
     def get_feasible_solution(self):
         """ 根据约束寻找到一个可行解 """
         import numpy as np
-        fsb_lst = np.zeros(len(self.variables))
+        fsb_lst = np.zeros(len(self.variables), dtype=int)
 
         t_machine = 0
         t_load = 0
