+++ janusq/application/chocoq/chocoq/problems/set_cover_problem.py
@@ -15,7 +15,7 @@ class SetCoverProblem(LcboModel):
     def get_feasible_solution(self):
         """ 根据约束寻找到一个可行解 """
         import numpy as np
-        fsb_lst = np.zeros(len(self.variables))
+        fsb_lst = np.zeros(len(self.variables), dtype=int)
         have_cover_set = set()
         for i in range(self.num_sets):
             fsb_lst[i] = 1
